Allow sending more than one message per program counter.
This is done by using a double-ended queue (a deque) to hold a list of
data received or deferreds to be triggered by incoming data. When data
arrives the deque is checked to see if there is a matching Deferred
that we can trigger. Otherwise the data is appended to the deque.
Likewise, when data is expected, the deque is checked for waiting
data. If no data is waiting, the Deferred is appended to the deque.

Data or Deferreds are appended to the right and popped from the left.