Changed broadcast state variables to be local to each broadcast.
Instead of having global tables for broadcast in general, we now use
local tables for each broadcast method call such that the tables can
be garbage collected when they are no longer needed. It is now also
unnecessary to store a table for each program counter, making the
local tables smaller than the previous global tables.